Fixed Logic Error - Instance Variables vs Class Variables
Variables declared in the scope of the entire class are updated for all instances of the entire class (i.e. similar to static variables in Java).

In order to have variables be unique to the instance of an object in Python, they must be declared within the scope of the constructor (__init__).

Changes were made reflecting this language-specific feature to the Day and Exercise classes, and tested in main.